CircularBuffer class moved to /utils repository
